backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Celery and Redis
celery[redis]>=5.3.6
redis>=5.0.1

# WebSocket
websockets>=12.0

# Existing services dependencies
requests>=2.31.0
groq>=0.11.0
python-telegram-bot>=21.0
pyyaml>=6.0.1

# YouTube API
google-api-python-client>=2.108.0
google-auth>=2.25.2
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Video generation
moviepy>=1.0.3
gtts>=2.5.0
pillow>=10.1.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0

